+++ app/modules/kit-library/kit-library-menu-item.php
@@ -3,13 +3,12 @@ namespace Elementor\App\Modules\KitLibrary;
 
 use Elementor\Core\Admin\Menu\Interfaces\Admin_Menu_Item;
 use Elementor\TemplateLibrary\Source_Local;
-use Elementor\Core\Admin\Menu\Interfaces\Admin_Menu_Item_Has_Position;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly.
 }
 
-class Kit_Library_Menu_Item implements Admin_Menu_Item, Admin_Menu_Item_Has_Position {
+class Kit_Library_Menu_Item implements Admin_Menu_Item {
 
 	public function is_visible() {
 		return true;
@@ -25,9 +24,5 @@ class Kit_Library_Menu_Item implements Admin_Menu_Item, Admin_Menu_Item_Has_Posi
 
 	public function get_capability() {
 		return 'manage_options';
-	}
-
-	public function get_position() {
-		return 30;
 	}
 }
