+++ core/role-manager/role-manager-menu-item.php
@@ -1,7 +1,6 @@
 <?php
 namespace Elementor\Core\RoleManager;
 
-use Elementor\Core\Admin\Menu\Interfaces\Admin_Menu_Item_Has_Position;
 use Elementor\Core\Admin\Menu\Interfaces\Admin_Menu_Item_With_Page;
 use Elementor\Settings;
 
@@ -9,7 +8,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly.
 }
 
-class Role_Manager_Menu_Item implements Admin_Menu_Item_With_Page, Admin_Menu_Item_Has_Position {
+class Role_Manager_Menu_Item implements Admin_Menu_Item_With_Page {
 
 	private $role_manager;
 
@@ -39,9 +38,5 @@ class Role_Manager_Menu_Item implements Admin_Menu_Item_With_Page, Admin_Menu_It
 
 	public function render() {
 		$this->role_manager->display_settings_page();
-	}
-
-	public function get_position() {
-		return 40;
 	}
 }
